requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
httpx==0.28.1
azure-identity==1.19.0
azure-cosmos==4.9.0
aiohttp==3.11.13
PyJWT[crypto]==2.10.1
pytest==8.3.5
pytest-asyncio==0.25.3
